Connect implementation of LEARNER-3582 with LEARNER-3251
JIRA link of both tickets:
- https://openedx.atlassian.net/browse/LEARNER-3582
- https://openedx.atlassian.net/browse/LEARNER-3251
